+++ src/sat/MinisatCore_prop.h
@@ -42,7 +42,7 @@ namespace BEEV
 
     void setVerbosity(int v);
 
-    int nVars();
+    unsigned long nVars();
 
     void printStats();
